Bug 280909: Syntax and bindings for variadic templates.

+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/pdom/dom/cpp/PDOMCPPAnnotation.java
@@ -108,7 +108,7 @@ class PDOMCPPAnnotation {
 	 * @param annotation Annotation containing visibility information.
 	 * @return The visibility component of the annotation.
 	 */
-	public static int getVisibility(byte annotation) {
+	public static int getVisibility(int annotation) {
 		return (annotation >> VISIBILITY_OFFSET) & VISIBILITY_MASK;
 	}
 }
